vm: fix some size_t/uint64_t issues (12)

+++ src/cdisasm/disasm.cpp
@@ -148,7 +148,7 @@ void parse_data_substream(Stream &input_stream, uint64_t start, bool no_numbers)
     }
 }
 
-void parse_symbol_substream(Stream &input_stream, size_t start, bool no_numbers)
+void parse_symbol_substream(Stream &input_stream, uint64_t start, bool no_numbers)
 {
     size_t       position = 0;
     StreamReader reader(input_stream, position);
